refactor: centralize live artifact assertions
