+++ test/integration/android-emulator-e2e/live-assertions.ts
@@ -1,33 +1,19 @@
 import assert from 'node:assert/strict';
-import fs from 'node:fs';
-
 import { PUBLIC_COMMANDS } from '../../../src/command-catalog.ts';
 import type { RawSnapshotNode } from '@agent-device/kernel/snapshot';
 import type { SnapshotDiffLine } from '../../../src/snapshot/snapshot-diff.ts';
-import { isPlayableVideo } from '../../../src/utils/video.ts';
 import {
   assertFilesDiffer,
   assertJsonContains,
+  assertMp4File,
+  assertNonEmptyFile,
   createLiveDeviceAssertions,
 } from '../live-device-e2e/assertions.ts';
 import type { CliJsonResult } from '../cli-json.ts';
 import type { AndroidEmulatorBehaviorId } from './behavior-coverage.ts';
 import { type LiveContext, runStep, verifyCommand } from './live-harness.ts';
 
-export { assertFilesDiffer, assertJsonContains };
-
-export function assertNonEmptyFile(filePath: string, name: string): void {
-  assert.ok(fs.statSync(filePath).size > 0, `${name} artifact is empty: ${filePath}`);
-}
-
-export async function assertMp4File(filePath: string): Promise<void> {
-  assertNonEmptyFile(filePath, 'recording');
-  assert.equal(
-    await isPlayableVideo(filePath),
-    true,
-    `recording is not a finalized playable video: ${filePath}`,
-  );
-}
+export { assertFilesDiffer, assertJsonContains, assertMp4File, assertNonEmptyFile };
 
 export const { assertElementText, assertWaitSelector, assertWaitText, capturePng } =
   createLiveDeviceAssertions<AndroidEmulatorBehaviorId, LiveContext>(
