+++ test/integration/live-device-e2e/assertions.ts
@@ -2,6 +2,7 @@ import assert from 'node:assert/strict';
 import fs from 'node:fs';
 
 import { assertPngFile } from '../provider-scenarios/assertions.ts';
+import { isPlayableVideo } from '../../../src/utils/video.ts';
 import type { CliJsonResult } from '../cli-json.ts';
 import type { LiveDeviceContext } from './runtime.ts';
 
@@ -64,4 +65,17 @@ export function assertJsonContains(result: CliJsonResult, expected: string, mess
 
 export function assertFilesDiffer(first: string, second: string, message: string): void {
   assert.notDeepEqual(fs.readFileSync(first), fs.readFileSync(second), message);
+}
+
+export function assertNonEmptyFile(filePath: string, name: string): void {
+  assert.ok(fs.statSync(filePath).size > 0, `${name} artifact is empty: ${filePath}`);
+}
+
+export async function assertMp4File(filePath: string): Promise<void> {
+  assertNonEmptyFile(filePath, 'recording');
+  assert.equal(
+    await isPlayableVideo(filePath),
+    true,
+    `recording is not a finalized playable video: ${filePath}`,
+  );
 }
